Fix: Inner cluster communication with keycloak causes issuer verification to fail
Due to the fact the backend communicates with the keycloak directly via the k8s
service and not the route the issuer verification did fail as it was not the
service url but the route url. This commit introduces new settings for the
issuer host and issuer url that is then the url from the route as this is
the url the frontend is communicating with.