Add coffee drinks delete route
This commit adds the first coffee drink route. Its the delete case as
this is the first case where the app should support an coffee rating
(precessor name of coffee drink) action without the dependency to
a coffee bean. Other routes can and will get added to the newly created
coffee drinks router